Add files via upload
This is an analysis of the Covid19 data using the Standard Query Language(SQL). It querried the number of cases recorder per day in each country, continent and the world at large. Also the number of deaths recorded thereto. Further the percentage of number cases and deaths with regards to the population was also querried, etc.